+++ RELEASENOTES.md
@@ -1,5 +1,6 @@
 # 2025.10.3
 - Update export to sql routine
+- Add import sql routine
 
 # 2025.10.2
 - Display amount of users in the overview